--- exercises/concept/array-loops/.meta/config.json
@@ -1,11 +1,0 @@
-{
-  "blurb": "TODO: add blurb for array-loops exercise",
-  "authors": ["rishiosaur"],
-  "contributors": ["SleeplessByte"],
-  "files": {
-    "solution": ["array-loops.js"],
-    "test": ["array-loops.spec.js"],
-    "exemplar": [".meta/exemplar.js"]
-  },
-  "forked_from": []
-}
