+++ exercises/concept/elyses-looping-enchantments/.docs/hints.md
