+++ exercises/concept/elyses-looping-enchantments/package.json
@@ -1,11 +1,12 @@
 {
-  "name": "@exercism/javascript-array-loops",
+  "name": "@exercism/javascript-elyses-looping-enchantments",
   "description": "Exercism concept exercise on array-loops",
   "author": "Derk-Jan Karrenbeld <derk-jan+github@karrenbeld.info>",
   "private": true,
   "repository": {
     "type": "git",
-    "url": "https://github.com/exercism/javascript"
+    "url": "https://github.com/exercism/javascript",
+    "directory": "languages/javascript/exercises/concept/elyses-looping-enchantments"
   },
   "devDependencies": {
     "@babel/cli": "^7.13.14",
